Fixes and refactoring
