+++ src/library.ts
@@ -1,7 +1,8 @@
-import { createStore, combineReducers, applyMiddleware } from "redux";
+import { combineReducers, applyMiddleware } from "redux";
 const LogRocket = require("ui/utils/logrocket").default;
 import { isDevelopment, skipTelemetry } from "ui/utils/environment";
 import { sanityCheckMiddleware } from "ui/utils/sanitize";
+const configureStore = require("devtools/client/debugger/src/actions/utils/create-store").default;
 import reducer from "ui/reducers/app";
 const Account = require("ui/components/Account").default;
 
@@ -12,6 +13,7 @@ export async function initialize() {
       : undefined
     : applyMiddleware(LogRocket.reduxMiddleware());
 
+  const createStore = configureStore();
   const store = createStore(combineReducers({ app: reducer }), {}, middleware);
 
   return { store, Page: Account };
