+++ src/protocol/enable-repaint.ts
@@ -1,0 +1,9 @@
+let enableRepaint = false;
+
+export function isRepaintEnabled() {
+  return enableRepaint;
+}
+
+export function updateEnableRepaint(_enableRepaint: boolean) {
+  enableRepaint = _enableRepaint;
+}
