+++ src/ui/components/shared/OnboardingModal/OnboardingModal.tsx
@@ -1,7 +1,6 @@
-import classNames from "classnames";
 import React, { Dispatch, SetStateAction, useState } from "react";
 import { connect, ConnectedProps } from "react-redux";
-import { actions } from "ui/actions";
+import * as actions from "ui/actions/app";
 import hooks from "ui/hooks";
 import { Nag } from "ui/hooks/users";
 import Modal from "../NewModal";
