+++ src/ui/components/shared/WorkspaceSettingsModal/WorkspaceMember.tsx
@@ -2,10 +2,9 @@ import React, { useState } from "react";
 import hooks from "ui/hooks";
 import { WorkspaceUser } from "ui/types";
 import PortalDropdown from "ui/components/shared/PortalDropdown";
-import useToken from "ui/utils/useToken";
 import { connect, ConnectedProps } from "react-redux";
-import { selectors } from "ui/reducers";
-import { actions } from "ui/actions";
+import * as selectors from "ui/reducers/app";
+import * as actions from "ui/actions/app";
 import { UIState } from "ui/state";
 import { NonRegisteredTeamMember } from "ui/hooks/invitations";
 const { prefs } = require("ui/utils/prefs");
